Python/requirements.txt
flask
flask-cors
gspread
requests
python-dotenv
gunicorn